Move module into compatibility package
