+++ src/dapla_metadata/datasets/core.py
@@ -17,12 +17,14 @@ from datadoc_model.all_optional.model import DataSetStatus
 
 from dapla_metadata._shared import config
 from dapla_metadata.dapla import user_info
-from dapla_metadata.datasets.dapla_dataset_path_info import DaplaDatasetPathInfo
-from dapla_metadata.datasets.dataset_parser import DatasetParser
-from dapla_metadata.datasets.model_backwards_compatibility import (
+from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
     is_metadata_in_container_structure,
 )
-from dapla_metadata.datasets.model_backwards_compatibility import upgrade_metadata
+from dapla_metadata.datasets.compatibility.model_backwards_compatibility import (
+    upgrade_metadata,
+)
+from dapla_metadata.datasets.dapla_dataset_path_info import DaplaDatasetPathInfo
+from dapla_metadata.datasets.dataset_parser import DatasetParser
 from dapla_metadata.datasets.model_validation import ValidateDatadocMetadata
 from dapla_metadata.datasets.statistic_subject_mapping import StatisticSubjectMapping
 from dapla_metadata.datasets.utility.constants import (
